configuring heroku postgres DB

+++ src/database/migrations/01_create_reservations.ts
@@ -4,7 +4,7 @@ import { Knex } from 'knex'
 export async function up(knex: Knex) {
     return knex.schema.createTable('reservations', table => {
         table.increments('id').primary();
-        table.integer('date').notNullable(); //Date of the reservation.
+        table.bigInteger('date').notNullable(); //Date of the reservation.
         table.string('period').notNullable(); //Periods are Morning, Afternoon, Night.
         table.string('staff').notNullable(); //Name of the user doing the reservation.
